Add files via upload
This database contains 34 attributes, 33 of which are linear valued and one of them is nominal. The differential diagnosis of erythemato-squamous diseases is a real problem in dermatology. They all share the clinical features of erythema and scaling, with very little differences.

The diseases in this group are:

1)Psoriasis

2)Seboreic dermatitis

3)Lichen planus

4)Pityriasis rosea

5)Chronic dermatitis

6)Pityriasis rubra pilaris

Usually a biopsy is necessary for the diagnosis but unfortunately these diseases share many histopathological features as well.

Another difficulty for the differential diagnosis is that a disease may show the features of another disease at the beginning stage and may have the characteristic features at the following stages.

Patients were first evaluated clinically with 12 features. Afterwards, skin samples were taken for the evaluation of 22 histopathological features.

The values of the histopathological features are determined by an analysis of the samples under a microscope.